+++ Source/ConformalDecals/ModuleConformalDecal.cs
@@ -482,9 +482,6 @@ namespace ConformalDecals {
                 // skip disabled renderers
                 if (renderer.gameObject.activeInHierarchy == false) continue;
                 
-                // skip transparentFX layer, which causes issues with Waterfall
-                if (renderer.gameObject.layer == 1) continue;
-
                 // skip blacklisted shaders
                 if (DecalConfig.IsBlacklisted(renderer.material.shader)) continue;